Setting up the entities to better fit CA.

+++ src/main/java/entity/City.java
@@ -14,7 +14,7 @@ public interface City {
      * Returns the temperature of the city.
      * @return the temperature of the city.
      */
-    float getTemperature();
+    double getTemperature();
 
     /**
      * Returns the condition of the city.
